updated service tests
